Add NativeAudioPlayer

+++ src/jvmMain/kotlin/cn/apisium/eim/api/AudioPlayer.kt
@@ -3,5 +3,5 @@ package cn.apisium.eim.api
 import cn.apisium.eim.api.processor.AudioProcessor
 
 abstract class AudioPlayer(val currentPosition: CurrentPosition, var processor: AudioProcessor? = null): AutoCloseable {
-    abstract fun open(sampleRate: Float, bufferSize: Int, bits: Int)
+    abstract fun open(sampleRate: Int, bufferSize: Int, bits: Int)
 }
